test: split prepare hook coverage into dedicated file
